Add route handling and tests for trial-level files
Additional route handling was needed for files nested within a trial
(versus at the study-level).  The FileController was modified to handle
trial-level file routes (`/files/:study/:trial/:file`) in addition to
the study-level file routes (`/studies/:study/files/:file`).

Note that the same handlers are used for both route patterns.  They now
just check whether a trial parameter (`.../:trial/...` in the route,
retrieved with `p.ByName("trial")`) is present or not.  For example,
if getting a trial-level file resource, the `Get` handler will modify
the resource ID when it gets a trial parameter so that it retrieves the
appropriate resource. The resource ID is changed from `/studies/files/:file`
to `/files/:study/:trial/:file`.  # Please enter the commit message for
your changes. Lines starting # with '#' will be ignored, and an empty
message aborts the commit.