Add logic and tests for trial and study deletion
We needed to handle deletion of trials and files contained within a study when
that study gets deleted.  A prior commit added logic for deleting trials and
study-level files owned by the study to-be-deleted.  This commit adds logic for
dealing with trial-level files.

This commit also adds logic for deleting trial-level files when the parent
trial is deleted.